+++ src/System.Data.SqlClient/tests/ManualTests/DataCommon/DataTestUtility.cs
@@ -223,5 +223,42 @@ namespace System.Data.SqlClient.ManualTesting.Tests
         {
             ExpectFailure<AggregateException, TException, TInnerException>(() => actionThatFails().Wait(), null, exceptionMessage, innerExceptionMessage, innerInnerExceptionMustBeNull);
         }
+
+        public static string GenerateTableName()
+        {
+            return string.Format("TEST_{0}{1}{2}", Environment.GetEnvironmentVariable("ComputerName"), Environment.TickCount, Guid.NewGuid()).Replace('-', '_');
+        }
+
+        public static void RunNonQuery(string connectionString, string sql)
+        {
+            using (SqlConnection connection = new SqlConnection(connectionString))
+            {
+                connection.Open();
+                using (SqlCommand command = connection.CreateCommand())
+                {
+                    command.CommandText = sql;
+                    command.ExecuteNonQuery();
+                }
+            }
+        }
+
+        public static DataTable RunQuery(string connectionString, string sql)
+        {
+            DataTable result = null;
+            using (SqlConnection connection = new SqlConnection(connectionString))
+            {
+                connection.Open();
+                using (SqlCommand command = connection.CreateCommand())
+                {
+                    command.CommandText = sql;
+                    using (SqlDataReader reader = command.ExecuteReader())
+                    {
+                        result = new DataTable();
+                        result.Load(reader);
+                    }
+                }
+            }
+            return result;
+        }
     }
 }
