--- src/System.Data.SqlClient/tests/ManualTests/SQL/TestHelper/TestHelper.cs
@@ -1,48 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace System.Data.SqlClient.ManualTesting.Tests
-{
-    public class TestHelper
-    {
-        public static string GenerateTableName()
-        {
-            return string.Format("TEST_{0}{1}{2}", Environment.GetEnvironmentVariable("ComputerName"), Environment.TickCount, Guid.NewGuid()).Replace('-', '_');
-        }
-
-        public static void RunNonQuery(string connectionString, string sql)
-        {
-            using (SqlConnection connection = new SqlConnection(connectionString))
-            {
-                connection.Open();
-                using (SqlCommand command = connection.CreateCommand())
-                {
-                    command.CommandText = sql;
-                    command.ExecuteNonQuery();
-                }
-            }
-        }
-
-        public static DataTable RunQuery(string connectionString, string sql)
-        {
-            DataTable result = null;
-            using (SqlConnection connection = new SqlConnection(connectionString))
-            {
-                connection.Open();
-                using (SqlCommand command = connection.CreateCommand())
-                {
-                    command.CommandText = sql;
-                    using (SqlDataReader reader = command.ExecuteReader())
-                    {
-                        result = new DataTable();
-                        result.Load(reader);
-                    }
-                }
-            }
-            return result;
-        }
-    }
-}
